Moved init check to shared state
