+++ ParraCore/Parra+Push.swift
@@ -24,21 +24,23 @@ public extension Parra {
 
         let tokenString = token.map { String(format: "%02.2hhx", $0) }.joined()
 
-        guard Parra.Initializer.isInitialized else {
+        Task {
+            await registerDevicePushTokenString(tokenString)
+        }
+    }
+
+    internal class func registerDevicePushTokenString(_ tokenString: String) async {
+        guard await ParraGlobalState.shared.isInitialized() else {
             parraLogWarn("Parra.registerDevicePushToken was called before Parra.initialize(). Make sure that you're calling Parra.initialize() before application.registerForRemoteNotifications()")
 
             // Still try to recover from this situation. Temporarily cache the token, which we can check for when
             // initialization does occur and proceed to upload it.
-            Task {
-                await PushTokenState.shared.setTemporaryPushToken(tokenString)
-            }
+            await ParraGlobalState.shared.setTemporaryPushToken(tokenString)
 
             return
         }
 
-        Task {
-            await uploadDevicePushToken(tokenString)
-        }
+        await uploadDevicePushToken(tokenString)
     }
 
     internal class func uploadDevicePushToken(_ token: String) async {
@@ -48,13 +50,13 @@ public extension Parra {
             parraLogTrace("Device push token successfully uploaded. Clearing cache.")
             // Token shouldn't be cached when it isn't absolutely necessary to recover from errors. If a new token
             // is issued, it will replace the old one.
-            await PushTokenState.shared.clearTemporaryPushToken()
+            await ParraGlobalState.shared.clearTemporaryPushToken()
         } catch let error {
             parraLogTrace("Device push token failed to upload. Caching token to try later.")
             // In the event that the upload failed, we'll cache the token. This will be overridden by any new token
             // on the next app launch, but in the event that we're able to retry the request later, we'll have this
             // one on hand.
-            await PushTokenState.shared.setTemporaryPushToken(token)
+            await ParraGlobalState.shared.setTemporaryPushToken(token)
 
             parraLogError("Error uploading push token to Parra API", error)
         }
